+++ cucaibafront/src/Pages/TablaHonorarios.jsx
@@ -10,13 +10,11 @@ import { useSelector, useDispatch } from "react-redux";
 import {
   getOperativos,
   getAgentes,
-  getModulos,
-  getHonorario,
-  postHonorario,
 } from "../Redux/Actions";
-import Swal from "sweetalert2";
+import { FormHonorario } from "../components/FormHonorario";
 
 const ExpandedComponent = () => {
+
   //Renderizando los agentes //
   const agentes = useSelector((state) => state.agentes);
   let dispatch = useDispatch();
@@ -31,78 +29,6 @@ const ExpandedComponent = () => {
     setAgente(primerArreglo);
   }, []);
 
-  //Renderizando los módulos
-
-  const modulos = useSelector((state) => state.modulos);
-
-  let arregloModulos = [];
-
-  if (modulos.length > 1) {
-    arregloModulos = modulos[1][0];
-  }
-  const [modulo, setModulo] = useState(arregloModulos);
-
-  useEffect(() => {
-    dispatch(getModulos());
-  }, []);
-
-  useEffect(() => {
-    setModulo(arregloModulos);
-  }, []);
-
-  //CREACION DE HONORARIO //
-  const [honorario, setHonorario] = useState({
-    operativo_id: "",
-    agente_id: "",
-    modulo_id: "",
-    fechaModif: "",
-  });
-
-  const handleOnSubmit = async (e) => {
-    e.preventDefault();
-
-    if (
-      honorario.operativo_id &&
-      honorario.agente_id &&
-      honorario.modulo_id &&
-      honorario.fechaModif
-    ) {
-      dispatch(
-        postHonorario(
-          honorario.operativo_id,
-          honorario.agente_id,
-          honorario.modulo_id,
-          honorario.fechaModif
-        ));
-
-      //  ALERT //
-      await Swal.fire({
-        position: "center",
-        icon: "success",
-        title: "El honorario se ha asignado correctamente al Agente",
-        showConfirmButton: false,
-        timer: 3000,
-      });
-      window.location.reload();
-
-      setHonorario({
-        operativo_id: "",
-        agente_id: "",
-        modulo_id: "",
-        fechaModif: "",
-      });
-    } else {
-      Swal.fire({
-        position: "center",
-        icon: "error",
-        title: "Por favor, completa todos los campos",
-        showConfirmButton: true,
-      });
-
-      // FIN ALTERT //
-    }
-  };
-
   //MOSTRANDO EL FORMULARIO DE CREACION //
   const [mostrarFormulario, setMostrarFormulario] = useState(false);
   function handleMostrarFormulario() {
@@ -114,9 +40,7 @@ const ExpandedComponent = () => {
   }
 
   //FIN FORMULARIO DE CREACION//
-  const [opcionSeleccionada, setOpcionSeleccionada] = useState("");
-  const [nombreOpcionSeleccionada, setNombreOpcionSeleccionada] = useState("");
-
+  
   return (
     <div className="p-3">
       <div className="d-flex align-items-center justify-content-between">
@@ -151,141 +75,7 @@ const ExpandedComponent = () => {
                 border: "1px solid gray",
               }}
             >
-              <form onSubmit={handleOnSubmit} className="row g-3 pt-4">
-                <div
-                  className="honorario"
-                  style={{
-                    display: "flex",
-                    justifyContent: "flex-start",
-                    marginLeft: 0,
-                    color: "#5DADE2",
-                  }}
-                >
-                  <h6>CREAR HONORARIO</h6>
-                </div>
-                <hr
-                  style={{
-                    margin: "10px 0",
-                    border: "none",
-                    borderBottom: "5px solid #5DADE2",
-                  }}
-                />
-                <label className="NroPD"> Proceso de Donación Número </label>
-
-                <input
-  type="hidden"
-  value={honorario.operativo_id}
-  name="operativo_id"
-/>
-
-                <div className="col-md-6">
-                  <label htmlFor="tipo">
-                    Seleccionar Agente
-                    <span
-                      style={{
-                        color: "red",
-                        marginLeft: "5px",
-                        fontSize: "20px",
-                      }}
-                    >
-                      *
-                    </span>
-                  </label>
-                  <select
-                    className="form-select form-select-md mb-3"
-                    aria-label=".form-select-lg example"
-                    name="agente_id"
-                    value={honorario.agente_id}
-                    onChange={(e) =>
-                      setHonorario({
-                        ...honorario,
-                        agente_id: Number(e.target.value),
-                      })
-                    }
-                    placeholder="Selecciona un tipo"
-                  >
-                    <option value="">Seleccionar</option>
-                    {primerArreglo.map((agente) => (
-                      <option key={agente.id} value={agente.id}>
-                        {agente.apellido + ", " + agente.nombre}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-                <div className="col-md-6">
-                  <label htmlFor="tipo">
-                    Función desempeñada
-                    <span
-                      style={{
-                        color: "red",
-                        marginLeft: "5px",
-                        fontSize: "20px",
-                      }}
-                    >
-                      *
-                    </span>
-                  </label>
-                  <select
-                    className="form-select form-select-md mb-3"
-                    aria-label=".form-select-lg example"
-                    name="modulo_id"
-                    value={opcionSeleccionada}
-                    onChange={(e) => {
-                      setHonorario({
-                        ...honorario,
-                        modulo_id: Number(e.target.value),
-                      });
-                      setOpcionSeleccionada(e.target.value);
-                      setNombreOpcionSeleccionada(
-                        e.target.options[e.target.selectedIndex].text
-                      );
-                    }}
-                    placeholder="Selecciona un tipo"
-                  >
-                    <option value="">Seleccionar</option>
-                    {arregloModulos.map((modulo) => (
-                      <option key={modulo.id} value={modulo.id}>
-                        {modulo.descripcion}
-                      </option>
-                    ))}
-                  </select>
-
-                  {opcionSeleccionada && (
-                    <div className="pt-4">
-                      <p>{nombreOpcionSeleccionada}</p>
-                      <p style={{fontWeight:"bold"}}>
-                        Valor: {arregloModulos[opcionSeleccionada - 1]?.valor}
-                      </p>
-                    </div>
-                  )}
-                </div>
-
-                <input
-  type="hidden"
-  value={honorario.fechaModif}
-  name="fechaModif"
-/>
-                <hr
-                  style={{
-                    marginTop: "3rem",
-                    border: "none",
-                    borderBottom: "5px solid #5DADE2",
-                  }}
-                />
-                <div className="d-flex justify-content-end">
-                  <button
-                    onClick={handleCerrarFormulario}
-                    type="submit"
-                    className="btn btn-outline-secondary pb-2"
-                    style={{ marginRight: "10px" }}
-                  >
-                    Cancelar
-                  </button>
-                  <button type="submit" className="btn btn-success pt-2">
-                    Crear Honorario
-                  </button>
-                </div>
-              </form>
+              <FormHonorario handleCerrarFormulario={handleCerrarFormulario} />
             </div>
           </div>
         )}
